refactor: html tag 속성 수정
